# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillUp LMS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all open projects (public)
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Post a new project (client only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project (client only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (client only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/bids:
    get:
      operationId: getProjectBids
      tags: [projects]
      summary: List bids on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bids list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BidWithFreelancer"
    post:
      operationId: submitBid
      tags: [projects]
      summary: Submit a bid on a project (freelancer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidInput"
      responses:
        "201":
          description: Bid submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
  /bids/{id}:
    patch:
      operationId: updateBid
      tags: [projects]
      summary: Accept or reject a bid (client only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidPatch"
      responses:
        "200":
          description: Updated bid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
components:
  schemas:
...
    Project:
      type: object
      required: [id, clientId, title, description, category, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skillsRequired:
          type: ["string", "null"]
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        bidCount:
          type: integer
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skillsRequired:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
    ProjectPatch:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skillsRequired:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
    Bid:
      type: object
      required: [id, projectId, freelancerId, amount, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        freelancerId:
          type: integer
        amount:
          type: number
        duration:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
    BidInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        duration:
          type: string
        coverLetter:
          type: string
    BidPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]
    BidWithFreelancer:
      type: object
      required: [id, projectId, freelancerId, amount, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        freelancerId:
          type: integer
        freelancerName:
          type: string
        freelancerEmail:
          type: string
        freelancerTitle:
          type: ["string", "null"]
        amount:
          type: number
        duration:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
...