# lib/api-zod/src/generated/types/bidWithFreelancer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillUp LMS Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { BidWithFreelancerStatus } from './bidWithFreelancerStatus';

export interface BidWithFreelancer {
  id: number;
  projectId: number;
  freelancerId: number;
  freelancerName?: string;
  freelancerEmail?: string;
  /** @nullable */
  freelancerTitle?: string | null;
  amount: number;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  status: BidWithFreelancerStatus;
  createdAt: string;
}