# lib/api-zod/src/generated/types/projectPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillUp LMS Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectPatchStatus } from './projectPatchStatus';

export interface ProjectPatch {
  title?: string;
  description?: string;
  category?: string;
  skillsRequired?: string;
  budgetMin?: number;
  budgetMax?: number;
  deadline?: string;
  status?: ProjectPatchStatus;
}